создает PointController
